Issue #5: Move state definition to its own file
